address PR comments for Nest Sensor
