+++ requirements_all.txt
@@ -89,6 +89,9 @@ https://github.com/bashwork/pymodbus/archive/d7fc4f1cc975631e0a9011390e8017f64b6
 # homeassistant.components.mqtt
 paho-mqtt==1.1
 
+# homeassistant.components.nest
+python-nest==2.6.0
+
 # homeassistant.components.notify.pushbullet
 pushbullet.py==0.9.0
 
@@ -183,9 +186,6 @@ heatmiserV3==0.9.1
 
 # homeassistant.components.thermostat.honeywell
 evohomeclient==0.2.4
-
-# homeassistant.components.thermostat.nest
-python-nest==2.6.0
 
 # homeassistant.components.thermostat.radiotherm
 radiotherm==1.2
